module name rename to InventoryBundleProduct
