--- app/code/Magento/InventoryBundle/README.md
@@ -1,1 +1,0 @@
-# InventoryBundle
