+++ app/code/Magento/InventoryBundleProduct/LICENSE_AFL.txt
