+++ app/code/Magento/InventoryBundleProduct/README.md
@@ -1,0 +1,1 @@
+# InventoryBundleProduct
