+++ app/code/Magento/InventoryBundleProduct/Test/Api/BundleProductShouldBeInStockOnCreateTest.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventoryBundle\Test\Api;
+namespace Magento\InventoryBundleProduct\Test\Api;
 
 use Magento\Catalog\Api\Data\ProductInterface;
 use Magento\Catalog\Model\Product\Type;
