+++ app/code/Magento/InventoryBundleProduct/Test/Integration/CatalogInventory/Api/StockRegistry/GetStockStatusBySkuOnDefaultStockTest.php
@@ -5,25 +5,19 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventoryBundle\Test\Integration\CatalogInventory\Api\StockRegistry;
+namespace Magento\InventoryBundleProduct\Test\Integration\CatalogInventory\Api\StockRegistry;
 
 use Magento\CatalogInventory\Api\StockRegistryInterface;
 use Magento\InventoryCatalogApi\Api\DefaultStockProviderInterface;
-use Magento\InventoryCatalogApi\Model\GetProductIdsBySkusInterface;
 use Magento\TestFramework\Helper\Bootstrap;
 use PHPUnit\Framework\TestCase;
 
-class GetProductStockStatusOnDefaultStockTest extends TestCase
+class GetStockStatusBySkuOnDefaultStockTest extends TestCase
 {
     /**
      * @var StockRegistryInterface
      */
     private $stockRegistry;
-
-    /**
-     * @var GetProductIdsBySkusInterface
-     */
-    private $getProductIdsBySkus;
 
     /**
      * @var DefaultStockProviderInterface
@@ -36,12 +30,11 @@ class GetProductStockStatusOnDefaultStockTest extends TestCase
     protected function setUp()
     {
         $this->stockRegistry = Bootstrap::getObjectManager()->get(StockRegistryInterface::class);
-        $this->getProductIdsBySkus = Bootstrap::getObjectManager()->get(GetProductIdsBySkusInterface::class);
         $this->defaultStockProvider = Bootstrap::getObjectManager()->get(DefaultStockProviderInterface::class);
     }
 
     /**
-     * @magentoDataFixture ../../../../app/code/Magento/InventoryBundle/Test/_files/default_stock_bundle_products.php
+     * @magentoDataFixture ../../../../app/code/Magento/InventoryBundleProduct/Test/_files/default_stock_bundle_products.php
      *
      * @dataProvider getStockDataProvider
      * @param string $sku
@@ -50,14 +43,14 @@ class GetProductStockStatusOnDefaultStockTest extends TestCase
      */
     public function testGetStatusIfScopeIdParameterIsNotPassed(string $sku, int $status): void
     {
-        $productId = $this->getProductIdsBySkus->execute([$sku])[$sku];
-        $productStockStatus = $this->stockRegistry->getProductStockStatus($productId);
+        $stockStatus = $this->stockRegistry->getStockStatusBySku($sku);
 
-        self::assertEquals($status, $productStockStatus);
+        self::assertEquals($status, $stockStatus->getStockStatus());
+        self::assertEquals(0, $stockStatus->getQty());
     }
 
     /**
-     * @magentoDataFixture ../../../../app/code/Magento/InventoryBundle/Test/_files/default_stock_bundle_products.php
+     * @magentoDataFixture ../../../../app/code/Magento/InventoryBundleProduct/Test/_files/default_stock_bundle_products.php
      *
      * @dataProvider getStockDataProvider
      * @param string $sku
@@ -66,13 +59,10 @@ class GetProductStockStatusOnDefaultStockTest extends TestCase
      */
     public function testGetStatusIfScopeIdParameterIsPassed(string $sku, int $status): void
     {
-        $productId = $this->getProductIdsBySkus->execute([$sku])[$sku];
-        $productStockStatus = $this->stockRegistry->getProductStockStatus(
-            $productId,
-            $this->defaultStockProvider->getId()
-        );
+        $stockStatus = $this->stockRegistry->getStockStatusBySku($sku, $this->defaultStockProvider->getId());
 
-        self::assertEquals($status, $productStockStatus);
+        self::assertEquals($status, $stockStatus->getStockStatus());
+        self::assertEquals(0, $stockStatus->getQty());
     }
 
     /**
