+++ app/code/Magento/InventoryBundleProduct/Test/_files/default_stock_bundle_products.php
