+++ app/code/Magento/InventoryBundleProduct/composer.json
@@ -19,7 +19,7 @@
             "registration.php"
         ],
         "psr-4": {
-            "Magento\\InventoryBundle\\": ""
+            "Magento\\InventoryBundleProduct\\": ""
         }
     }
 }
