+++ app/code/Magento/InventoryBundleProduct/etc/module.xml
@@ -7,5 +7,5 @@
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Magento_InventoryBundle" setup_version="1.0.0" />
+    <module name="Magento_InventoryBundleProduct" setup_version="1.0.0" />
 </config>
